# openapi.yaml
openapi: 3.1.0
info:
  title: Növénydiagnosztikai API
  version: 1.0.0
servers:
  - url: https://plant-diagnosis-1.onrender.com

paths:
  /:
    get:
      operationId: rootGet
      summary: Gyors ellenőrzés (root)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required: [status]

  /health:
    get:
      operationId: healthGet
      summary: Egészség ellenőrzés
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required: [status]

  /identify:
    post:
      operationId: identifyPlant
      summary: Növény azonosítása képfájlból (multipart/form-data)
      parameters:
        - in: query
          name: organs
          schema:
            type: string
            default: leaf
          description: "Példa: leaf / flower / fruit / bark (több is lehet: leaf,flower)"
        - in: query
          name: project
          schema:
            type: string
            default: weurope
          description: "PlantNet project, pl. weurope vagy all"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
            encoding:
              image:
                contentType: image/jpeg, image/png, image/webp
      responses:
        "200":
          description: Sikeres azonosítás
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifyResponse"
        "422":
          description: Hibás kérés
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Külső szolgáltatás hiba (PlantNet)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /identify_b64:
    post:
      operationId: identifyPlantB64
      summary: Növény azonosítása base64 képből (JSON) – csak ha tényleg base64-et küldesz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_base64]
              properties:
                image_base64:
                  type: string
                  description: "Base64 string vagy data URL (data:image/jpeg;base64,...)"
                organs:
                  type: string
                  default: leaf
                project:
                  type: string
                  default: weurope
      responses:
        "200":
          description: Sikeres azonosítás
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifyResponse"
        "422":
          description: Hibás kérés (pl. nem érvényes base64)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Külső szolgáltatás hiba (PlantNet)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /diseases/identify:
    post:
      operationId: identifyDisease
      summary: Betegség/kártevő azonosítása képfájlból (multipart/form-data)
      parameters:
        - in: query
          name: project
          schema:
            type: string
            default: weurope
          description: "PlantNet project (ha releváns)"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
            encoding:
              image:
                contentType: image/jpeg, image/png, image/webp
      responses:
        "200":
          description: Sikeres azonosítás
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifyResponse"
        "422":
          description: Hibás kérés
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Külső szolgáltatás hiba (PlantNet)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /diseases/identify_b64:
    post:
      operationId: identifyDiseaseB64
      summary: Betegség/kártevő azonosítása base64 képből (JSON) – csak ha tényleg base64-et küldesz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_base64]
              properties:
                image_base64:
                  type: string
                  description: "Base64 string vagy data URL (data:image/jpeg;base64,...)"
                project:
                  type: string
                  default: weurope
      responses:
        "200":
          description: Sikeres azonosítás
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifyResponse"
        "422":
          description: Hibás kérés (pl. nem érvényes base64)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Külső szolgáltatás hiba (PlantNet)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /diagnose:
    post:
      operationId: diagnose
      summary: Kombinált diagnózis (növény + betegség/kártevő) képfájlból
      parameters:
        - in: query
          name: organs
          schema:
            type: string
            default: leaf
          description: "Példa: leaf / flower / fruit / bark (több is lehet: leaf,flower)"
        - in: query
          name: project
          schema:
            type: string
            default: weurope
          description: "PlantNet project, pl. weurope vagy all"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
            encoding:
              image:
                contentType: image/jpeg, image/png, image/webp
      responses:
        "200":
          description: Sikeres kombinált diagnózis
          content:
            application/json:
              schema:
                type: object
                required: [plant, diseaseOrPest, summary]
                properties:
                  plant:
                    $ref: "#/components/schemas/IdentifyResponse"
                  diseaseOrPest:
                    $ref: "#/components/schemas/IdentifyResponse"
                  summary:
                    type: object
                    properties:
                      bestPlant:
                        type: string
                      plantScore:
                        type: number
                      bestIssue:
                        type: string
                      issueScore:
                        type: number
                      project:
                        type: string
                      organs:
                        type: array
                        items:
                          type: string
        "422":
          description: Hibás kérés
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Külső szolgáltatás hiba (PlantNet)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    IdentifyResponse:
      type: object
      required: [bestMatch, confidence, topMatches]
      properties:
        bestMatch:
          type: string
        confidence:
          type: object
          required: [top1_score, level]
          properties:
            top1_score:
              type: number
            level:
              type: string
        topMatches:
          type: array
          items:
            type: object
            required: [name, score]
            properties:
              name:
                type: string
              score:
                type: number
        meta:
          type: object
          additionalProperties: true

    ErrorResponse:
      type: object
      additionalProperties: true
      properties:
        detail:
          description: Hiba részlete (szöveg vagy objektum)
          anyOf:
            - type: string
            - type: object